requirements.txt
asgiref==3.7.2
bitcoin==1.1.42
boto3==1.29.3
certifi==2023.7.22
charset-normalizer==3.2.0
distlib==0.3.7
dj-database-url==2.1.0
Django==4.2.4
django-cors-headers==4.2.0
django-extensions==3.2.3
django-filter==23.3
django-heroku==0.3.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
django-storages==1.14.2
Faker==19.11.0
filelock==3.12.2
gunicorn==21.2.0
idna==3.4
packaging==23.1
platformdirs==3.10.0
plaid-python==16.0.0
psycopg2==2.9.7
pybitcointools==1.1.15
PyJWT==2.8.0
python-dotenv==1.0.0
pytz==2023.3
requests==2.31.0
sqlparse==0.4.4
stripe==5.5.0
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.0.4
virtualenv==20.24.2
whitenoise==6.5.0
